==> 02-plan_kayak_trip_WEBSCRAPING.ipynb <==
# Planning my next vacay ☀️ Part 2

Get info about hotels in specific cities with Scrapy via <a href="https://www.booking.com" target="_blank">www.booking.com</a> .

Scape info like:

* hotel name 
* url to hotel's booking.com page, 
* coordinates: latitude and longitude,
* score given by the website users,
* text description of the hotel

Then excute this script for various cities from the list in part 1.
Save results from each city in their own file.

import os
import logging

import scrapy
from scrapy.crawler import CrawlerProcess

destination_name = 'La Rochelle'

class Hotels(scrapy.Spider):
    # define spider
    name = "hotels"

    # define starting URL
    start_urls = ['https://www.booking.com/index.fr.html']

    # Parse function for login
    def parse(self, response):
        # use FormRequest to login
        return scrapy.FormRequest.from_response(
            response,
            formdata={'ss': destination_name},
            callback=self.after_search
        )

    # Callback used after login
    def after_search(self, response):
        
        hotels = response.css('.sr_item')

        for h in hotels:
            yield {
                'name': h.css('.sr-hotel__name::text').get(),
                'url': "https://www.booking.com" + h.css('.hotel_name_link').attrib["href"],
                'coords': h.css('.sr_card_address_line a').attrib["data-coords"],
                'score': h.css('.bui-review-score__badge::text').get(),
                'description': h.css('.hotel_desc::text').get()
                
            }
        
        
        # Define NEXT button and store in next_page
        try:
            next_page = response.css('a.paging-next').attrib["href"]
        except KeyError:
            logging.info('No next page. Terminating crawling process.')
        else:
            yield response.follow(next_page, callback=self.after_search)


filename = "2_hotels_" + destination_name.replace(" ", "-") + ".json"

if filename in os.listdir('res/'):
        os.remove('res/' + filename)

process = CrawlerProcess(settings = {
    'USER_AGENT': 'Chrome/84.0 (compatible; MSIE 7.0; Windows NT 5.1)',
    'LOG_LEVEL': logging.INFO,
    "FEEDS": {
        'res/' + filename: {"format": "json"},
    }
})

process.crawl(Hotels)
process.start()
