==> 04-plan_kayak_trip_Amazon_RDS.ipynb <==
# Planning my next vacay ☀️ Part 4

Put dataset in postgreSQL database on Amazon RDS. Use SQL syntax to make requests and decide where to go for vacation.

!pip install psycopg2-binary

Read `.csv` files containing information about the cities and hotels:

import pandas as pd
from sqlalchemy import create_engine

cities = pd.read_csv('res/1_destinations.csv')
hotels = pd.read_csv('res/3_hotels.csv')

cities.head()

hotels.head()

Create and configure RDS instance on AWS. 

Use SQLAlchemy to create an engine that's connected to the remote database:

# Change YOUR_USERNAME, YOUR_PASSWORD and YOUR_HOST with you credentials
engine = create_engine("postgresql+psycopg2://YOUR_USERNAME:YOUR_PASSWORD@YOUR_HOST/postgres", echo=True)

4. Create two tables, `cities` and `hotels`, in remote database with info saved in `.csv` files from previous sections.

cities.to_sql("cities", engine)
hotels.to_sql("hotels", engine)

Use PGAdmin to explore the database

Use SQL syntax to make requests that will get the list of the 30 best places to go for vacay.

conn = engine.connect()

statement = "SELECT city_name, main_weather, day_temperature, hotels.name, description, score, url \
             FROM cities \
             INNER JOIN hotels \
             ON cities.city_id = hotels.city_id \
             WHERE rank=1 AND score IS NOT NULL \
             ORDER BY score DESC \
             LIMIT 30"

best_places = pd.read_sql(statement, conn)

display(best_places)
